Refatora exceptions e lógica de usuário

+++ src/main/java/com/br/unifil/vendas_analytics/vendas_analytics/service/CategoriaService.java
@@ -3,12 +3,15 @@ package com.br.unifil.vendas_analytics.vendas_analytics.service;
 import com.br.unifil.vendas_analytics.vendas_analytics.config.UsuarioAutenticadoDto;
 import com.br.unifil.vendas_analytics.vendas_analytics.model.Categoria;
 import com.br.unifil.vendas_analytics.vendas_analytics.repository.CategoriaRepository;
-import com.br.unifil.vendas_analytics.vendas_analytics.validation.ValidacaoException;
+import com.br.unifil.vendas_analytics.vendas_analytics.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Collections;
 import java.util.List;
+
+import static com.br.unifil.vendas_analytics.vendas_analytics.ExceptionMessage.ProdutoExceptionMessage.CATEGORIA_NAO_ENCONTRADA;
+import static com.br.unifil.vendas_analytics.vendas_analytics.ExceptionMessage.ProdutoExceptionMessage.CATEGORIA_VINCULADA_PRODUTO;
 
 @Service
 public class CategoriaService {
@@ -18,8 +21,8 @@ public class CategoriaService {
     @Autowired
     private UsuarioService usuarioService;
 
-    private static final ValidacaoException CATEGORIA_NOT_FOUND_EXCEPTION =
-        new ValidacaoException("Categoria não encontrado");
+    @Autowired
+    private ProdutoRepository produtoRepository;
 
     public void save(Categoria categoria) {
         categoria.setUsuarioCadastro(usuarioService.getUsuarioLogado().getId());
@@ -28,7 +31,7 @@ public class CategoriaService {
 
     public Categoria buscarUm(Integer id) {
         return categoriaRepository.findByIdAndUsuarioCadastroIn(id, getIdsDoUsuarioProprietario())
-            .orElseThrow(() -> CATEGORIA_NOT_FOUND_EXCEPTION);
+            .orElseThrow(CATEGORIA_NAO_ENCONTRADA::getException);
     }
 
     public List<Categoria> buscarTodos() {
@@ -37,7 +40,10 @@ public class CategoriaService {
 
     public void remover(Integer id) {
         Categoria categoria = categoriaRepository.findByIdAndUsuarioCadastroIn(id, getIdsDoUsuarioProprietario())
-            .orElseThrow(() -> CATEGORIA_NOT_FOUND_EXCEPTION);
+            .orElseThrow(CATEGORIA_NAO_ENCONTRADA::getException);
+        if (produtoRepository.existsByCategoria(categoria)) {
+            throw CATEGORIA_VINCULADA_PRODUTO.getException();
+        }
         categoriaRepository.delete(categoria);
     }
